Allow tapping on an unpopulated grid row without generating an error
The code that determines if a row that has been tapped had been
previously selected would fail if an empty (unpopulated) row was
tapped. This is because it was referencing the position of the
tapped row number in the data array, which didn't exist. The result
was a console message when run in a browser, and program termination
when run as a Windows app (both were observed when running our mobile
client on those Cordova platforms).

Fixed by checking that the tapped row number is a valid grid row.