7008. Added ability for user to reorder Tab Layout tabs in runtime and in design mode via drag-and-drop.
Fixed a bug with tab layouts where if a tab layout had the "active tab" property set and was inside a lazy-loaded layout, then rendering that layout would cause the screen to submit and not be able to select the tab or accordion section.

Tab Layout (and future pui.layout.Template subclasses) now support a .render function that is called after all property setters, allowing simpler rendering logic.